+++ Source/Core/Core/CMakeLists.txt
@@ -59,6 +59,8 @@ add_library(core
   Core.h
   CoreTiming.cpp
   CoreTiming.h
+  Debugger/CodeTrace.cpp
+  Debugger/CodeTrace.h
   Debugger/Debugger_SymbolMap.cpp
   Debugger/Debugger_SymbolMap.h
   Debugger/Dump.cpp
