+++ Source/Core/Core/Debugger/CodeTrace.cpp
@@ -1,7 +1,7 @@
 // Copyright 2022 Dolphin Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "Common/Debug/CodeTrace.h"
+#include "Core/Debugger/CodeTrace.h"
 
 #include <algorithm>
 #include <chrono>
