+++ Source/Core/Core/Debugger/CodeTrace.h
