+++ Source/Core/DolphinQt/Debugger/CodeViewWidget.cpp
@@ -24,10 +24,10 @@
 #include <QWheelEvent>
 
 #include "Common/Assert.h"
-#include "Common/Debug/CodeTrace.h"
 #include "Common/GekkoDisassembler.h"
 #include "Common/StringUtil.h"
 #include "Core/Core.h"
+#include "Core/Debugger/CodeTrace.h"
 #include "Core/Debugger/PPCDebugInterface.h"
 #include "Core/PowerPC/MMU.h"
 #include "Core/PowerPC/PPCAnalyst.h"
